feat: deleting images uploaded (local and s3)

+++ api/src/services/Post/DeletePost/DeletePostUseCases.ts
@@ -1,6 +1,7 @@
 import { APIError } from './../../../errors/APIError';
 import { IPost } from '../../../models/Post';
 import { IPostRepository } from './../../../repositories/IPostRepository';
+import { DeleteImage } from '../../../providers/DeleteImage';
 
 export class DeletePostUseCases {
 
@@ -9,6 +10,12 @@ export class DeletePostUseCases {
   ) { }
 
   async execute({ id }: { id: string }): Promise<void> {
+    const post = await this.postsRepository.listById(id);
+    if (!post) {
+      return;
+    }
+
+    await DeleteImage(post.key);
     await this.postsRepository.delete(id);
   }
